Iteration 04 - ETL to GCP Buckets/stage_04_vm_etl_big_query/requirements_debian_linux.txt
apache-beam==2.32.0
attrs==21.2.0
avro-python3==1.9.2.1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
fastavro==1.4.4
future==0.18.2
google-api-core==2.0.1
google-api-python-client==2.22.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-bigquery==2.28.0
google-cloud-bigquery-storage==2.9.0
google-cloud-core==2.0.0
google-cloud-storage==1.42.2
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.40.0
hdfs==2.6.0
httplib2==0.19.1
idna==3.2
libcst==0.3.21
mypy-extensions==0.4.3
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.1.1
orjson==3.6.3
packaging==21.0
pandas==1.3.3
pandas-gbq==0.15.0
pkg_resources==0.0.0
proto-plus==1.19.2
protobuf==3.18.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.2.0
pydot==1.4.2
pymongo==3.12.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tqdm==4.62.3
typing-extensions==3.7.4.3
typing-inspect==0.7.1
uritemplate==3.0.1
urllib3==1.26.6
